requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3
polars==0.20.2

# Technical Analysis
# ta-lib==0.4.28
pandas-ta==0.3.14b0

# Machine Learning & AI
torch==2.1.2
transformers==4.36.2
stable-baselines3==2.2.1
optuna==3.5.0
mlflow==2.9.2
scikit-learn==1.3.2

# Trading & Market Data
MetaTrader5==5.0.45
ccxt==4.1.64

# Database & Storage
sqlalchemy==2.0.25
# sqlite3  # Built-in with Python
redis==5.0.1
pyodbc==5.0.1  # For SQL Server if needed

# Async & Performance
asyncio  # Built-in with Python
aiohttp==3.9.1
# uvloop==0.19.0  # Linux only - skip on Windows

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
pywin32==306  # Windows specific
pydantic==2.5.2
click==8.1.7

# Monitoring & Logging
flask==3.0.0
fastapi==0.108.0
uvicorn==0.25.0
python-telegram-bot==20.7

# Math & Statistics
scipy==1.11.4
statsmodels==0.14.1

# File Processing
openpyxl==3.1.2
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Visualization (optional)
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0